requirements.txt
python-telegram-bot==22.5
deep-translator==1.11.4
python-dotenv
requests


